refactor: added test section for carousel component
