+++ app/components/Carousel/CarouselPagination/CarouselPagination.scss
@@ -5,6 +5,7 @@
   align-items: center;
   gap: 16px;
   & > span {
+    cursor: pointer;
     display: inline-block;
     width: 16px;
     height: 16px;
